Moved project to postgres + docker-compose improvements (#2)
* Moved project to postgres + docker-compose improvements

- Added new requirements for working with postgres

- Changed name of "downloader.py" to "worker.py" as eventually it will do
much more than just download.

- Using docker-compose secrets where supported. These are pulled into the
Python code as environment variables. Given the fact that they are already
stored in plaintext in and on the system, the fact they are environment
variables is not at this time a big deal. Would like to improve it later though
with possible encrypted credentials.